Interleaved Strings
http://www.practice.geeksforgeeks.org/problem-page.php?pid=700353

Given three strings A, B and C your task is to complete the function
isInterleave which returns true if C is an interleaving of A and B else
returns false. C is said to be interleaving A and B, if it contains all
characters of A and B and order of all characters in individual strings
is preserved.