Q2.5
You have two numbers represented by a linked list, where each node
contains a
single digit. The digits are stored in reverse order, such that the Ts
digit is at the
head of the list. Write a function that adds the two numbers and
returns the sum
as a linked list.