Q4.8
You have two very large binary trees: Tl, with millions of nodes, and
T2, with
hundreds of nodes. Create an algorithm to decide ifT2 is a subtree of
Tl.
A tree T2 is a subtree of Tl if there exists a node n in Tl such that
the subtree of
n is identical to T2. That is, if you cut off the tree at node n, the
two trees would
be identical.